src/quality-tester/requirements.txt
Flask
gunicorn
pyyaml